CRUD operations working perfectly

+++ AccountSystemRepo/src/main/java/za/ac/nwu/ac/repo/dto/Member_Repo.java
@@ -4,7 +4,7 @@ import org.springframework.stereotype.Repository;
 import za.ac.nwu.ac.domain.dto.Member;
 
 @Repository
-public interface Member_Repo extends JpaRepository<Member, Integer>
+public interface Member_Repo extends JpaRepository<Member, Long>
 {
 
 }
